pytools/cloud_src/requirements.txt
google-cloud-vision>=2.2.0
Pillow